requirements.txt
mpmath==1.3.0
pillow==10.4.0
sympy==1.13.2